delete from cloud if the checkbox is checked

+++ frappe_s3_attachment/controller.py
@@ -12,6 +12,8 @@ import magic
 import botocore
 import frappe
 
+from botocore.exceptions import ClientError
+
 from werkzeug.wrappers import Response
 
 
@@ -257,30 +259,28 @@ def migrate_existing_files():
 
 def delete_from_cloud(doc, method):
     """Delete file from s3"""
-    from botocore.exceptions import ClientError
-
     s3_settings_doc = frappe.get_doc(
-            'S3 File Attachment',
-            'S3 File Attachment',
-        )
-
-    S3 = boto3.resource('s3', region_name=s3_settings_doc.region_name)
-    S3_CLIENT = boto3.client(
-        's3',
-        aws_access_key_id=s3_settings_doc.aws_key,
-        aws_secret_access_key=s3_settings_doc.aws_secret,
-        region_name=s3_settings_doc.region_name,
+        'S3 File Attachment',
+        'S3 File Attachment',
     )
-    BUCKET = s3_settings_doc.bucket_name
 
+    if s3_settings_doc.delete_file_from_cloud:
 
-    try:
-        S3_CLIENT.delete_object(
-            Bucket=s3_settings_doc.bucket_name,
-            Key=doc.content_hash
+        S3 = boto3.resource('s3', region_name=s3_settings_doc.region_name)
+        S3_CLIENT = boto3.client(
+            's3',
+            aws_access_key_id=s3_settings_doc.aws_key,
+            aws_secret_access_key=s3_settings_doc.aws_secret,
+            region_name=s3_settings_doc.region_name,
         )
-    except ClientError as e:
-        frappe.throw("Access denied")
+
+        try:
+            S3_CLIENT.delete_object(
+                Bucket=s3_settings_doc.bucket_name,
+                Key=doc.content_hash
+            )
+        except ClientError as e:
+            frappe.throw("Access denied: Could not delete file")
 
 
 @frappe.whitelist()
